chore(restructure): move s3_uploader -> openshorts/integrations/s3.py
Phase 1 step 1: relocate the leaf module with the fewest reverse imports
(only app.py imports from s3_uploader). Adds a re-export shim at the old
path so existing `from s3_uploader import ...` keeps working through the
restructure. Tests stay 62/62 green.
